Move encodeobjects to modules
