+++ packages/stargate/src/index.ts
@@ -36,24 +36,6 @@ export {
   isAminoMsgWithdrawValidatorCommission,
 } from "./aminomsgs";
 export { AminoConverter, AminoTypes, AminoTypesOptions } from "./aminotypes";
-export {
-  isMsgDelegateEncodeObject,
-  isMsgDepositEncodeObject,
-  isMsgSendEncodeObject,
-  isMsgSubmitProposalEncodeObject,
-  isMsgTransferEncodeObject,
-  isMsgUndelegateEncodeObject,
-  isMsgVoteEncodeObject,
-  isMsgWithdrawDelegatorRewardEncodeObject,
-  MsgDelegateEncodeObject,
-  MsgDepositEncodeObject,
-  MsgSendEncodeObject,
-  MsgSubmitProposalEncodeObject,
-  MsgTransferEncodeObject,
-  MsgUndelegateEncodeObject,
-  MsgVoteEncodeObject,
-  MsgWithdrawDelegatorRewardEncodeObject,
-} from "./encodeobjects";
 export { calculateFee, GasPrice } from "./fee";
 export * as logs from "./logs";
 export {
@@ -64,8 +46,24 @@ export {
   GovParamsType,
   GovProposalId,
   IbcExtension,
+  isMsgDelegateEncodeObject,
+  isMsgDepositEncodeObject,
+  isMsgSendEncodeObject,
+  isMsgSubmitProposalEncodeObject,
+  isMsgTransferEncodeObject,
+  isMsgUndelegateEncodeObject,
+  isMsgVoteEncodeObject,
+  isMsgWithdrawDelegatorRewardEncodeObject,
   MintExtension,
   MintParams,
+  MsgDelegateEncodeObject,
+  MsgDepositEncodeObject,
+  MsgSendEncodeObject,
+  MsgSubmitProposalEncodeObject,
+  MsgTransferEncodeObject,
+  MsgUndelegateEncodeObject,
+  MsgVoteEncodeObject,
+  MsgWithdrawDelegatorRewardEncodeObject,
   setupAuthExtension,
   setupBankExtension,
   setupDistributionExtension,
