+++ packages/stargate/src/modules/distribution/messages.ts
@@ -1,4 +1,4 @@
-import { GeneratedType } from "@cosmjs/proto-signing";
+import { EncodeObject, GeneratedType } from "@cosmjs/proto-signing";
 import {
   MsgFundCommunityPool,
   MsgSetWithdrawAddress,
@@ -12,3 +12,17 @@ export const distributionTypes: ReadonlyArray<[string, GeneratedType]> = [
   ["/cosmos.distribution.v1beta1.MsgWithdrawDelegatorReward", MsgWithdrawDelegatorReward],
   ["/cosmos.distribution.v1beta1.MsgWithdrawValidatorCommission", MsgWithdrawValidatorCommission],
 ];
+
+export interface MsgWithdrawDelegatorRewardEncodeObject extends EncodeObject {
+  readonly typeUrl: "/cosmos.distribution.v1beta1.MsgWithdrawDelegatorReward";
+  readonly value: Partial<MsgWithdrawDelegatorReward>;
+}
+
+export function isMsgWithdrawDelegatorRewardEncodeObject(
+  object: EncodeObject,
+): object is MsgWithdrawDelegatorRewardEncodeObject {
+  return (
+    (object as MsgWithdrawDelegatorRewardEncodeObject).typeUrl ===
+    "/cosmos.distribution.v1beta1.MsgWithdrawDelegatorReward"
+  );
+}
