+++ packages/stargate/src/modules/ibc/messages.ts
@@ -1,4 +1,4 @@
-import { GeneratedType } from "@cosmjs/proto-signing";
+import { EncodeObject, GeneratedType } from "@cosmjs/proto-signing";
 import { MsgTransfer } from "cosmjs-types/ibc/applications/transfer/v1/tx";
 import {
   MsgAcknowledgement,
@@ -46,3 +46,12 @@ export const ibcTypes: ReadonlyArray<[string, GeneratedType]> = [
   ["/ibc.core.connection.v1.MsgConnectionOpenInit", MsgConnectionOpenInit],
   ["/ibc.core.connection.v1.MsgConnectionOpenTry", MsgConnectionOpenTry],
 ];
+
+export interface MsgTransferEncodeObject extends EncodeObject {
+  readonly typeUrl: "/ibc.applications.transfer.v1.MsgTransfer";
+  readonly value: Partial<MsgTransfer>;
+}
+
+export function isMsgTransferEncodeObject(object: EncodeObject): object is MsgTransferEncodeObject {
+  return (object as MsgTransferEncodeObject).typeUrl === "/ibc.applications.transfer.v1.MsgTransfer";
+}
