+++ packages/stargate/src/modules/index.ts
@@ -1,16 +1,34 @@
 export { AuthExtension, setupAuthExtension } from "./auth/queries";
 export { authzTypes } from "./authz/messages";
-export { bankTypes } from "./bank/messages";
+export { bankTypes, isMsgSendEncodeObject, MsgSendEncodeObject } from "./bank/messages";
 export { BankExtension, setupBankExtension } from "./bank/queries";
-export { distributionTypes } from "./distribution/messages";
+export {
+  distributionTypes,
+  isMsgWithdrawDelegatorRewardEncodeObject,
+  MsgWithdrawDelegatorRewardEncodeObject,
+} from "./distribution/messages";
 export { DistributionExtension, setupDistributionExtension } from "./distribution/queries";
 export { feegrantTypes } from "./feegrant/messages";
-export { govTypes } from "./gov/messages";
+export {
+  govTypes,
+  isMsgDepositEncodeObject,
+  isMsgSubmitProposalEncodeObject,
+  isMsgVoteEncodeObject,
+  MsgDepositEncodeObject,
+  MsgSubmitProposalEncodeObject,
+  MsgVoteEncodeObject,
+} from "./gov/messages";
 export { GovExtension, GovParamsType, GovProposalId, setupGovExtension } from "./gov/queries";
-export { ibcTypes } from "./ibc/messages";
+export { ibcTypes, isMsgTransferEncodeObject, MsgTransferEncodeObject } from "./ibc/messages";
 export { IbcExtension, setupIbcExtension } from "./ibc/queries";
 export { MintExtension, MintParams, setupMintExtension } from "./mint/queries";
 export { setupSlashingExtension, SlashingExtension } from "./slashing/queries";
-export { stakingTypes } from "./staking/messages";
+export {
+  isMsgDelegateEncodeObject,
+  isMsgUndelegateEncodeObject,
+  MsgDelegateEncodeObject,
+  MsgUndelegateEncodeObject,
+  stakingTypes,
+} from "./staking/messages";
 export { setupStakingExtension, StakingExtension } from "./staking/queries";
 export { setupTxExtension, TxExtension } from "./tx/queries";
