+++ packages/stargate/src/signingstargateclient.spec.ts
@@ -11,7 +11,7 @@ import protobuf from "protobufjs/minimal";
 
 import { AminoMsgDelegate } from "./aminomsgs";
 import { AminoTypes } from "./aminotypes";
-import { MsgDelegateEncodeObject, MsgSendEncodeObject } from "./encodeobjects";
+import { MsgDelegateEncodeObject, MsgSendEncodeObject } from "./modules";
 import { PrivateSigningStargateClient, SigningStargateClient } from "./signingstargateclient";
 import { assertIsDeliverTxFailure, assertIsDeliverTxSuccess, isDeliverTxFailure } from "./stargateclient";
 import {
